remove fullscreen login
