--- WalletWasabi.Fluent/Views/Login/FullScreenLoginView.axaml
@@ -1,78 +1,0 @@
-<UserControl xmlns="https://github.com/avaloniaui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:c="clr-namespace:WalletWasabi.Fluent.Controls"
-             xmlns:login="clr-namespace:WalletWasabi.Fluent.ViewModels.Login"
-             xmlns:navBar="clr-namespace:WalletWasabi.Fluent.ViewModels.NavBar"
-             xmlns:conv="using:WalletWasabi.Fluent.Converters"
-             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             x:DataType="login:FullScreenLoginViewModel"
-             x:CompileBindings="True"
-             x:Class="WalletWasabi.Fluent.Views.Login.FullScreenLoginView">
-  <c:ContentArea Title="Login"
-                 Caption="Select the wallet and enter its password"
-                 IsBusy="{Binding IsBusy}"
-                 ScrollViewer.VerticalScrollBarVisibility="Disabled">
-    <Panel>
-
-
-      <c:LoginControl Margin="0 0 0 40"/>
-
-      <ComboBox SelectedItem="{Binding SelectedWallet}"
-                Items="{Binding Wallets}"
-                HorizontalAlignment="Right"
-                VerticalAlignment="Bottom"
-                Background="Transparent"
-                Width="200">
-        <ComboBox.Styles>
-          <Style Selector="ComboBox /template/ Border#Background">
-            <Setter Property="CornerRadius" Value="0" />
-          </Style>
-          <Style Selector="ComboBox /template/ Border#HighlightBackground">
-            <Setter Property="CornerRadius" Value="0" />
-          </Style>
-        </ComboBox.Styles>
-        <ComboBox.Resources>
-          <CornerRadius x:Key="OverlayCornerRadius">0</CornerRadius>
-          <Thickness x:Key="ComboBoxDropdownBorderThickness">0</Thickness>
-        </ComboBox.Resources>
-        <ComboBox.ItemTemplate>
-          <DataTemplate>
-            <StackPanel DataContext="{Binding Wallet}" Orientation="Horizontal">
-              <Image Source="{Binding KeyManager, Converter={x:Static conv:WalletIconConverter.KeyManagerToImage}}" Margin="0 0 10 0" Width="20" Height="20"/>
-              <TextBlock Text="{Binding WalletName}" VerticalAlignment="Center" />
-            </StackPanel>
-          </DataTemplate>
-        </ComboBox.ItemTemplate>
-      </ComboBox>
-
-      <StackPanel Orientation="Horizontal"
-                  HorizontalAlignment="Left"
-                  VerticalAlignment="Bottom">
-        <StackPanel.Styles>
-          <Style Selector="PathIcon"
-                 x:DataType="navBar:NavBarItemViewModel">
-            <Setter Property="ToolTip.Tip" Value="{Binding Title}"/>
-            <Setter Property="Width" Value="25"/>
-            <Setter Property="Height" Value="25"/>
-            <Setter Property="Data" Value="{Binding IconName, Converter={x:Static conv:NavBarIconConverter.Instance}}"/>
-          </Style>
-
-          <Style Selector="Button.bottom">
-            <Setter Property="Command" Value="{Binding NavigateToCommand}"/>
-            <Setter Property="Padding" Value="10"/>
-          </Style>
-        </StackPanel.Styles>
-
-        <Button Classes="invisible bottom" CommandParameter="{Binding ActionCenter}">
-          <PathIcon DataContext="{Binding ActionCenter}"/>
-        </Button>
-
-        <Button Classes="invisible bottom" CommandParameter="{Binding AddWallet}">
-          <PathIcon DataContext="{Binding AddWallet}"/>
-        </Button>
-      </StackPanel>
-    </Panel>
-  </c:ContentArea>
-</UserControl>
