--- WalletWasabi.Fluent/Views/Login/FullScreenLoginView.axaml.cs
@@ -1,18 +1,0 @@
-using Avalonia.Controls;
-using Avalonia.Markup.Xaml;
-
-namespace WalletWasabi.Fluent.Views.Login
-{
-	public class FullScreenLoginView : UserControl
-	{
-		public FullScreenLoginView()
-		{
-			InitializeComponent();
-		}
-
-		private void InitializeComponent()
-		{
-			AvaloniaXamlLoader.Load(this);
-		}
-	}
-}
